+++ src/etsTargetFunction.cpp
@@ -165,8 +165,9 @@ void EtsTargetFunction::eval(const double* p_par, int p_par_length) {
 
 	for(int i=0; i <= p*this->y.size(); i++) state.push_back(0);
 
+	int nb = 0;
 	etscalc(&this->y[0], &this->n, &this->state[0], &this->m, &this->errortype, &this->trendtype, &this->seasontype,
-			&this->alpha, &this->beta, &this->gamma, &this->phi, &this->e[0], &this->fits[0], &this->lik, &this->amse[0], &this->nmse);
+			&this->alpha, &this->beta, &this->gamma, &this->phi, &this->e[0], &this->fits[0], &this->lik, &this->amse[0], &this->nmse, &nb);
 
 	// Avoid perfect fits
 	if (this->lik < -1e10) this->lik = -1e10;