+++ NEWS.md
@@ -2,8 +2,8 @@
 
 ## New features
 
-* `ARIMA()`, `RW()`, `NAIVE()` and `SNAIVE()` models can now be extended with
-  new data via `stream()`, which updates the fitted model without
+* `ARIMA()`, `ETS()`, `RW()`, `NAIVE()` and `SNAIVE()` models can now be
+  extended with new data via `stream()`, which updates the fitted model without
   re-estimating its coefficients by only needing to process the new values.
 
 ## Improvements